agenttests/localtools/requirements.txt
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-ai-agentserver-core[tracing]==2.0.0a20260331006
azure-ai-agentserver-invocations==1.0.0a20260331006
langgraph>=0.4.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
